fix: socperf interface LimitRequest adapt int64_t int type

+++ services/native/src/thermal_action/action/action_soc/action_gpu.cpp
@@ -93,7 +93,7 @@ void ActionGpu::Execute()
 int32_t ActionGpu::GpuRequest(uint32_t freq)
 {
     std::vector<int32_t> tags;
-    std::vector<int32_t> configs;
+    std::vector<int64_t> configs;
 
     if (!g_service->GetSimulationXml()) {
         tags.push_back(LIM_GPU_ID);
